# docs/engineering/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Forma.PublicApi | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:7225/"
    }
  ],
  "paths": {
...
    "/api/exercises/{id}": {
      "get": {
        "tags": [
          "Exercises"
        ],
        "summary": "Also backs ADR-006 Rule 1 (Exercise-existence check at Workout create/edit) once consumed by training-planning-service. Hand-authored against the live controller (Forma.PublicApi/Controllers/v1/ExercisesController.cs) — regenerate from a running instance once available to confirm exact shape.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfExerciseQueryModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorResponse": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ApiResponse": {
        "required": [
          "success",
          "successMessage",
          "statusCode",
          "errors"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "successMessage": {
            "type": "string",
            "nullable": true
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            },
            "nullable": true
          }
        }
      },
...
      "ApiResponseOfExerciseQueryModel": {
        "required": [
          "result",
          "success",
          "successMessage",
          "statusCode",
          "errors"
        ],
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/ExerciseQueryModel"
          },
          "success": {
            "type": "boolean"
          },
          "successMessage": {
            "type": "string",
            "nullable": true
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            },
            "nullable": true
          }
        }
      },
...
      "ExerciseQueryModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "muscleGroups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MuscleGroup"
            },
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "ownerId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
      "MuscleGroup": {
        "enum": [
          "Chest",
          "Back",
          "Legs",
          "Shoulders",
          "Arms",
          "Core",
          "FullBody"
        ]
      },
...
    }
  },
...
}